UGENE-8153 refactor dialog
