+++ src/plugins/enzymes/src/EnzymesPlugin.cpp
@@ -47,8 +47,9 @@
 #include "DigestSequenceDialog.h"
 #include "EnzymesQuery.h"
 #include "EnzymesTests.h"
-#include "FindEnzymesDialog.h"
 #include "FindEnzymesTask.h"
+#include "enzymes_dialog/EnzymesSelectorWidget.h"
+#include "enzymes_dialog/FindEnzymesDialog.h"
 #include "insert/InsertEnzymeDialog.h"
 
 const QString CREATE_PCR_PRODUCT_ACTION_NAME = "Create PCR product";
